feat(wasm): the browser playground — engine to wasm32, chunked OPFS loader, static site
The four lib crates compile to wasm32-unknown-unknown unchanged except
for named seams: asupersync pinned to its wasm-browser profile for
wasm targets only; the KernelTeam returns None on wasm (serial is the
only correct path); CodecCheckpoint hydrates serially on wasm; and
byte-based constructors (MappedFile/SafetensorsFile/MappedFttsq
from_bytes, *_mapped/_from_file loader entry points) replace the
filesystem, with the artifact digest-verified in wasm before any
tensor is read.

crates/ftts-wasm exposes the pipeline to JS: presets (the seven CLI
voices, same bytes), WasmEngine::synthesize (production sampler,
artifact-native int8 hydration so hot weights never widen to f32),
enroll (lazy speaker encoder), and bench_frame_kernels — the Spike-B
receipt: 258 ms per 80 ms frame for the talker+micro schedule,
single-thread SIMD128 in Node (~0.31x RT before codec; the page says
0.2-0.3x and calls the native CLI faster).

site/ is the playground (franken_markdown's demo look): chunked
resumable Range downloads through a same-origin Pages Function proxy
(GitHub release assets send no CORS headers), OPFS cache with pinned
SHA-256 verification, worker-hosted engine, mic-clone that never sends
audio anywhere, WebAudio playback, WAV download/share, COOP/COEP
headers ready for the future threads build.

+++ crates/ftts-wasm/src/lib.rs
@@ -21,7 +21,9 @@ use ftts_model_qwen::prompt::{CloneMode, PromptMode};
 use ftts_model_qwen::sampler::SamplingMode;
 use ftts_model_qwen::speaker::{Encoder as SpeakerEncoder, log_mel_from_24khz_pcm};
 use ftts_model_qwen::talker::TalkerConfig;
-use ftts_model_qwen::tokenizer::{QwenTokenizer, TokenizerFiles};
+use ftts_model_qwen::tokenizer::QwenTokenizer;
+#[cfg(not(unix))]
+use ftts_model_qwen::tokenizer::TokenizerFiles;
 use std::path::Path;
 use std::sync::Arc;
 use wasm_bindgen::prelude::*;
@@ -381,7 +383,7 @@ pub fn bench_frame_kernels(rounds: u32) -> String {
         matrices.push((QuantizedMatrix { data, scales, n, k }, calls));
     }
     let x: Vec<i8> = (0..8192)
-        .map(|i| (((i * 29 + 5) % 255) as i32 - 127) as i8)
+        .map(|i| (((i * 29 + 5) % 255) - 127) as i8)
         .collect();
     let mut out = vec![0.0f32; 8192];
